big project update + added comments
after a lot of tries in different solving algorithms I have decided to try and implement the dancing links x (DLX) algorithm into my sudoku solver.

added -
DLXAlgorithm: this module has the DLX implementation algorithm itself and some helper methods.
MakeCoverMatrix: this module creates the cover matrix for the exact cover program of the DLX algorithm.
MakeDoubleLinkedList: this module creates a double linked list from the cover matrix to enable the DLX algorithm the most efficient way to work
Node:  a class of a node in the double linked list
ColumnNode: a class of a Node who is the column in the double linked list
ColumnID: a class that saves the information of the columns

changes -
changed the entire project to fit with the dancing algorithm